+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bypass spatial join when using only one patch to speed-up `make_transcript_patches` (#362)
 
 ### Fixed
+- Update `attrs` on disk when updating the latest Sopa boundaries (#363)
 - Ensure shapes transformation is always 2D (#359)
 
 ## [2.1.9] - 2025-11-14